refactor: extract paragraph parsing
